Quick start section
My first attempt to use TinyLogger was mostly positive, but I realized I was not using it properly because the same message was appearing many times. This is because I was using `addFileLoggerNamed:` in my code, not realizing I had to remove the previous logger(s). So, I suggest adding a quick start to get the basic use down.